requirements.txt
phidata 
openai 
duckduckgo-search
yfinance
fastapi
uvicorn
sqlalchemy
pgvector
psycopg[binary]
pypdf
typer
psycopg2